+++ controllers/om/api/admin.go
@@ -3,12 +3,37 @@ package api
 import (
 	"encoding/json"
 	"fmt"
+	"net/http"
 	"net/url"
 
 	"github.com/10gen/ops-manager-kubernetes/controllers/om/apierror"
+	"github.com/10gen/ops-manager-kubernetes/pkg/util/versionutil"
 
 	"github.com/10gen/ops-manager-kubernetes/controllers/om/backup"
 )
+
+type Key struct {
+	Description string              `json:"desc"`
+	ID          string              `json:"id"`
+	Links       []map[string]string `json:"links"`
+	PrivateKey  string              `json:"privateKey"`
+	PublicKey   string              `json:"publicKey"`
+	Roles       []map[string]string `json:"roles"`
+}
+
+type GlobalApiKeyRequest struct {
+	Description string   `json:"desc"`
+	Roles       []string `json:"roles"`
+}
+
+type KeyResponse struct {
+	ApiKeys []Key `json:"results"`
+}
+
+type Whitelist struct {
+	CidrBlock   string `json:"cidrBlock"`
+	Description string `json:"description"`
+}
 
 // Admin (imported as 'api.Admin') is the client to all "administrator" related operations with Ops Manager
 // which do not relate to specific groups (that's why it's different from 'om.Connection'). The only state expected
@@ -56,8 +81,17 @@ type Admin interface {
 	// DeleteBlockStoreConfig removes the Block store by its ID
 	DeleteBlockStoreConfig(id string) error
 
-	// ReadFileSystemStoreConfig reads the FileSystemSnapshot sotre by its ID
+	// ReadFileSystemStoreConfig reads the FileSystemSnapshot store by its ID
 	ReadFileSystemStoreConfigs() ([]backup.DataStoreConfig, error)
+
+	// ReadGlobalAPIKeys reads the global API Keys in Ops Manager
+	ReadGlobalAPIKeys() ([]Key, error)
+
+	// CreateGlobalAPIKey creates a new Global API Key in Ops Manager
+	CreateGlobalAPIKey(description string) (Key, error)
+
+	// ReadOpsManagerVersion read the version returned in the Header
+	ReadOpsManagerVersion() (versionutil.OpsManagerVersion, error)
 }
 
 // AdminProvider is a function which returns an instance of Admin interface initialized with connection parameters.
@@ -79,7 +113,7 @@ func NewOmAdmin(baseUrl, user, publicApiKey string) Admin {
 }
 
 func (a *DefaultOmAdmin) ReadDaemonConfig(hostName, headDbDir string) (backup.DaemonConfig, error) {
-	ans, apiErr := a.get("admin/backup/daemon/configs/%s/%s", hostName, url.QueryEscape(headDbDir))
+	ans, _, apiErr := a.get("admin/backup/daemon/configs/%s/%s", hostName, url.QueryEscape(headDbDir))
 	if apiErr != nil {
 		return backup.DaemonConfig{}, apiErr
 	}
@@ -94,7 +128,7 @@ func (a *DefaultOmAdmin) ReadDaemonConfig(hostName, headDbDir string) (backup.Da
 func (a *DefaultOmAdmin) CreateDaemonConfig(hostName, headDbDir string) error {
 	config := backup.NewDaemonConfig(hostName, headDbDir)
 	// dev note, for creation we don't specify the second path parameter (head db) - it's used only during update
-	_, err := a.put("admin/backup/daemon/configs/%s", config, hostName)
+	_, _, err := a.put("admin/backup/daemon/configs/%s", config, hostName)
 	if err != nil {
 		return err
 	}
@@ -105,7 +139,7 @@ func (a *DefaultOmAdmin) CreateDaemonConfig(hostName, headDbDir string) error {
 // Some assumption: while the API returns the paginated source we don't handle it to make api simpler (quite unprobable
 // to have 500+ configs)
 func (a *DefaultOmAdmin) ReadOplogStoreConfigs() ([]backup.DataStoreConfig, error) {
-	res, err := a.get("admin/backup/oplog/mongoConfigs/")
+	res, _, err := a.get("admin/backup/oplog/mongoConfigs/")
 	if err != nil {
 		return nil, err
 	}
@@ -120,13 +154,13 @@ func (a *DefaultOmAdmin) ReadOplogStoreConfigs() ([]backup.DataStoreConfig, erro
 
 // CreateOplogStoreConfig creates an oplog store in Ops Manager
 func (a *DefaultOmAdmin) CreateOplogStoreConfig(config backup.DataStoreConfig) error {
-	_, err := a.post("admin/backup/oplog/mongoConfigs/", config)
+	_, _, err := a.post("admin/backup/oplog/mongoConfigs/", config)
 	return err
 }
 
 // UpdateOplogStoreConfig updates an oplog store in Ops Manager
 func (a *DefaultOmAdmin) UpdateOplogStoreConfig(config backup.DataStoreConfig) error {
-	_, err := a.put("admin/backup/oplog/mongoConfigs/%s", config, config.Id)
+	_, _, err := a.put("admin/backup/oplog/mongoConfigs/%s", config, config.Id)
 	return err
 }
 
@@ -139,7 +173,7 @@ func (a *DefaultOmAdmin) DeleteOplogStoreConfig(id string) error {
 // Some assumption: while the API returns the paginated source we don't handle it to make api simpler (quite unprobable
 // to have 500+ configs)
 func (a *DefaultOmAdmin) ReadBlockStoreConfigs() ([]backup.DataStoreConfig, error) {
-	res, err := a.get("admin/backup/snapshot/mongoConfigs/")
+	res, _, err := a.get("admin/backup/snapshot/mongoConfigs/")
 	if err != nil {
 		return nil, err
 	}
@@ -154,13 +188,13 @@ func (a *DefaultOmAdmin) ReadBlockStoreConfigs() ([]backup.DataStoreConfig, erro
 
 // CreateBlockStoreConfig creates an Block store in Ops Manager
 func (a *DefaultOmAdmin) CreateBlockStoreConfig(config backup.DataStoreConfig) error {
-	_, err := a.post("admin/backup/snapshot/mongoConfigs/", config)
+	_, _, err := a.post("admin/backup/snapshot/mongoConfigs/", config)
 	return err
 }
 
 // UpdateBlockStoreConfig updates an Block store in Ops Manager
 func (a *DefaultOmAdmin) UpdateBlockStoreConfig(config backup.DataStoreConfig) error {
-	_, err := a.put("admin/backup/snapshot/mongoConfigs/%s", config, config.Id)
+	_, _, err := a.put("admin/backup/snapshot/mongoConfigs/%s", config, config.Id)
 	return err
 }
 
@@ -171,17 +205,17 @@ func (a *DefaultOmAdmin) DeleteBlockStoreConfig(id string) error {
 
 // S3 related methods
 func (a *DefaultOmAdmin) CreateS3Config(s3Config backup.S3Config) error {
-	_, err := a.post("admin/backup/snapshot/s3Configs", s3Config)
+	_, _, err := a.post("admin/backup/snapshot/s3Configs", s3Config)
 	return err
 }
 
 func (a *DefaultOmAdmin) UpdateS3Config(s3Config backup.S3Config) error {
-	_, err := a.put("admin/backup/snapshot/s3Configs/%s", s3Config, s3Config.Id)
+	_, _, err := a.put("admin/backup/snapshot/s3Configs/%s", s3Config, s3Config.Id)
 	return err
 }
 
 func (a *DefaultOmAdmin) ReadS3Configs() ([]backup.S3Config, error) {
-	res, err := a.get("admin/backup/snapshot/s3Configs")
+	res, _, err := a.get("admin/backup/snapshot/s3Configs")
 	if err != nil {
 		return nil, apierror.New(err)
 	}
@@ -198,7 +232,7 @@ func (a *DefaultOmAdmin) DeleteS3Config(id string) error {
 }
 
 func (a *DefaultOmAdmin) ReadFileSystemStoreConfigs() ([]backup.DataStoreConfig, error) {
-	res, err := a.get("admin/backup/snapshot/fileSystemConfigs/")
+	res, _, err := a.get("admin/backup/snapshot/fileSystemConfigs/")
 	if err != nil {
 		return nil, err
 	}
@@ -211,17 +245,80 @@ func (a *DefaultOmAdmin) ReadFileSystemStoreConfigs() ([]backup.DataStoreConfig,
 	return dataStoreConfigResponse.DataStoreConfigs, nil
 }
 
+// ReadGlobalAPIKeys reads the global API Keys in Ops Manager
+func (a *DefaultOmAdmin) ReadGlobalAPIKeys() ([]Key, error) {
+	res, _, err := a.get("admin/apiKeys")
+	if err != nil {
+		return nil, err
+	}
+
+	apiKeyResponse := &KeyResponse{}
+	if err = json.Unmarshal(res, apiKeyResponse); err != nil {
+		return nil, apierror.New(err)
+	}
+
+	return apiKeyResponse.ApiKeys, nil
+}
+
+//addWhitelistEntryIfItDoesntExist adds a whitelist through OM API. If it already exists, in just retun
+func (a *DefaultOmAdmin) addWhitelistEntryIfItDoesntExist(cidrBlock string, description string) error {
+	_, _, err := a.post("admin/whitelist", Whitelist{
+		CidrBlock:   cidrBlock,
+		Description: description,
+	})
+	if apierror.NewNonNil(err).ErrorCode == apierror.DuplicateWhitelistEntry {
+		return err
+	}
+	return nil
+}
+
+// CreateGlobalAPIKey creates a new Global API Key in Ops Manager.
+func (a *DefaultOmAdmin) CreateGlobalAPIKey(description string) (Key, error) {
+	if err := a.addWhitelistEntryIfItDoesntExist("0.0.0.0/1", description); err != nil {
+		return Key{}, err
+	}
+	if err := a.addWhitelistEntryIfItDoesntExist("128.0.0.0/1", description); err != nil {
+		return Key{}, err
+	}
+
+	newKeyBytes, _, err := a.post("admin/apiKeys", GlobalApiKeyRequest{
+		Description: description,
+		Roles:       []string{"GLOBAL_OWNER"},
+	})
+	if err != nil {
+		return Key{}, err
+	}
+
+	apiKey := &Key{}
+	if err := json.Unmarshal(newKeyBytes, apiKey); err != nil {
+		return Key{}, err
+	}
+	return *apiKey, nil
+
+}
+
+// ReadOpsManagerVersion read the version returned in the Header.
+func (a *DefaultOmAdmin) ReadOpsManagerVersion() (versionutil.OpsManagerVersion, error) {
+	_, header, err := a.get("")
+	if err != nil {
+		return versionutil.OpsManagerVersion{}, err
+	}
+	return versionutil.OpsManagerVersion{
+		VersionString: versionutil.GetVersionFromOpsManagerApiHeader(header.Get("X-MongoDB-Service-Version")),
+	}, nil
+}
+
 //********************************** Private methods *******************************************************************
 
-func (a *DefaultOmAdmin) get(path string, params ...interface{}) ([]byte, error) {
+func (a *DefaultOmAdmin) get(path string, params ...interface{}) ([]byte, http.Header, error) {
 	return a.httpVerb("GET", path, nil, params...)
 }
 
-func (a *DefaultOmAdmin) put(path string, v interface{}, params ...interface{}) ([]byte, error) {
+func (a *DefaultOmAdmin) put(path string, v interface{}, params ...interface{}) ([]byte, http.Header, error) {
 	return a.httpVerb("PUT", path, v, params...)
 }
 
-func (a *DefaultOmAdmin) post(path string, v interface{}, params ...interface{}) ([]byte, error) {
+func (a *DefaultOmAdmin) post(path string, v interface{}, params ...interface{}) ([]byte, http.Header, error) {
 	return a.httpVerb("POST", path, v, params...)
 }
 
@@ -230,19 +327,18 @@ func (a *DefaultOmAdmin) post(path string, v interface{}, params ...interface{})
 //}
 
 func (a *DefaultOmAdmin) delete(path string, params ...interface{}) error {
-	_, err := a.httpVerb("DELETE", path, nil, params...)
+	_, _, err := a.httpVerb("DELETE", path, nil, params...)
 	return err
 }
 
-func (a *DefaultOmAdmin) httpVerb(method, path string, v interface{}, params ...interface{}) ([]byte, error) {
+func (a *DefaultOmAdmin) httpVerb(method, path string, v interface{}, params ...interface{}) ([]byte, http.Header, error) {
 	client, err := NewHTTPClient(OptionDigestAuth(a.User, a.PublicAPIKey), OptionSkipVerify)
 	if err != nil {
-		return nil, apierror.New(err)
+		return nil, nil, apierror.New(err)
 	}
 
 	path = fmt.Sprintf("/api/public/v1.0/%s", path)
 	path = fmt.Sprintf(path, params...)
 
-	response, _, apiErr := client.Request(method, a.BaseURL, path, v)
-	return response, apiErr
+	return client.Request(method, a.BaseURL, path, v)
 }