+++ controllers/om/apierror/api_error.go
@@ -12,6 +12,7 @@ const (
 	BackupDaemonConfigNotFound = "DAEMON_MACHINE_CONFIG_NOT_FOUND"
 	UserAlreadyExists          = "USER_ALREADY_EXISTS"
 	S3ConfigNotFound           = "S3_SNAPSHOT_CONFIG_NOT_FOUND"
+	DuplicateWhitelistEntry    = "DUPLICATE_GLOBAL_WHITELIST_ENTRY"
 )
 
 // Error is the error extension that contains the details of OM error if OM returned the error. This allows the