+++ docker/mongodb-enterprise-tests/tests/opsmanager/om_ops_manager_https.py
@@ -96,7 +96,7 @@ def test_appdb_enable_tls(
     ops_manager.update()
     ops_manager.appdb_status().assert_abandons_phase(Phase.Running, timeout=60)
     ops_manager.appdb_status().assert_reaches_phase(Phase.Running, timeout=400)
-    ops_manager.om_status().assert_reaches_phase(Phase.Running, timeout=400)
+    ops_manager.om_status().assert_reaches_phase(Phase.Running, timeout=800)
 
 
 @mark.e2e_om_ops_manager_https_enabled